# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /imports/{id}/production-movement:
    get:
      operationId: getImportProductionMovement
      tags: [imports]
      summary: Cutting output (mark-level) + net balance delta for consecutive imports
      description: >
        Walks backwards from the given import (up to 7 predecessors) and, for
        each consecutive pair, computes: (1) TLT Cutting output — mark-level
        balance drawdown between snapshots (marks that left C + marks still at C
        whose weight decreased); (2) net balance delta per activity (current
        minus previous import balance weight in MT). FG delta is derived from
        parseSummary.fgWipByJob. Only consecutive pairs are compared — gaps
        (elapsed days > 1) are flagged but still returned. Purely additive and
        advisory — never changes parsing, activity, dedup, ageing, or any
        warning/milestone/dispatch state.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cutting output + net balance delta for consecutive import pairs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionMovementResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /imports/{id}/contractor-movement:
    get:
      operationId: getImportContractorMovement
      tags: [imports]
      summary: Per-contractor balance movement for consecutive imports (mark-level)
      description: >
        Walks backwards from the given import (up to 7 predecessors) and, for
        each consecutive pair, computes five mark-level measures per contractor
        (TLT / orderNature = "Structure" only, no project code forward-fill):
        Produced (weight actually drawn down — same-contractor reductions + marks
        that fully left WIP); Received (weight reassigned in from another
        contractor); Released (weight reassigned out to another contractor);
        New Intake (new marks attributed to this contractor); Net Change
        (overall balance delta). Purely additive and advisory — never changes
        parsing, activity, dedup, ageing, or any warning/milestone/dispatch state.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Per-contractor movement for consecutive import pairs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractorBalanceMovementResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ProductionMovementResponse:
      type: object
      description: >
        Cutting output (mark-level TLT) and net balance delta per activity for
        consecutive import pairs. Days are in chronological order (oldest first).
      properties:
        days:
          type: array
          items:
            $ref: "#/components/schemas/ProductionMovementDay"
      required:
        - days
    ProductionMovementDay:
      type: object
      description: >
        One consecutive import pair. cuttingOutputMt is the sum of TLT balance
        weight drawn down from marks that were at activity C in the previous
        import (left C entirely, or still at C with reduced weight). netBalance
        maps activity code to MT delta (curr − prev); negative = clearing,
        positive = accumulating. isGap=true when the two imports are more than
        one calendar day apart. gated=true when either import in the pair
        pre-dates per-row Type/Status storage; in that case cuttingOutputMt=0
        and netBalance={} — the pair cannot be computed.
      properties:
        importId:
          type: integer
          description: ID of the current (newer) import.
        dayKey:
          type: string
          description: YYYY-MM-DD of the current import (reportDate, else createdAt UTC date).
        dayLabel:
          type: string
          description: Human-readable label, e.g. "20 Jul" or "18-20 Jul (2d)".
        prevImportId:
          type: integer
          description: ID of the previous (older) import.
        prevDayKey:
          type: string
          description: YYYY-MM-DD of the previous import.
        isGap:
          type: boolean
          description: True when the two imports are more than one calendar day apart.
        elapsedDays:
          type: integer
          description: Calendar days between the two import dates.
        cuttingOutputMt:
          type: number
          description: >
            Mark-level TLT cutting output in MT. Sum of balance weight of marks
            that left C entirely plus the weight reduction of marks still at C.
            Excludes marks entering C (intake) and weight increases (corrections).
        cuttingMarksLeft:
          type: integer
          description: Count of TLT marks that were at C in prev and are no longer at C in curr.
        cuttingMarksReduced:
          type: integer
          description: Count of TLT marks still at C in both imports whose balance weight decreased.
        netBalance:
          type: object
          description: >
            Net balance delta in MT per activity (curr balance − prev balance).
            Negative = material leaving (good); positive = material accumulating (potential bottleneck).
            Key "FG" is derived from parseSummary fgWipByJob. Empty ({}) when gated=true.
          additionalProperties:
            type: number
        gated:
          type: boolean
          description: >-
            True when either import in this pair pre-dates per-row Type/Status storage.
            Movement cannot be computed; cuttingOutputMt and netBalance are zeroed.
        gatedReason:
          type: string
          description: Names the specific unusable import when gated is true.
      required:
        - importId
        - dayKey
        - dayLabel
        - prevImportId
        - prevDayKey
        - isGap
        - elapsedDays
        - cuttingOutputMt
        - cuttingMarksLeft
        - cuttingMarksReduced
        - netBalance
    ContractorBalanceMovementResponse:
      type: object
      description: Per-contractor balance movement across consecutive import pairs.
      properties:
        days:
          type: array
          items:
            $ref: "#/components/schemas/ContractorBalanceMovementDay"
      required:
        - days
    ContractorBalanceMeasures:
      type: object
      description: Five balance-movement measures for one contractor in one consecutive import pair.
      properties:
        produced:
          type: number
          description: >
            MT actually drawn down by this contractor — same-contractor weight
            reductions plus marks that fully left WIP. Excludes reassignments,
            new intake, and weight increases (corrections). Always >= 0.
        received:
          type: number
          description: MT of marks reassigned IN to this contractor from another.
        released:
          type: number
          description: MT of marks reassigned OUT from this contractor to another.
        newIntake:
          type: number
          description: MT of marks appearing for the first time in the current snapshot, attributed to this contractor.
        netChange:
          type: number
          description: >
            Overall balance delta in MT (curr total − prev total for this
            contractor). Positive = workload growing; negative = clearing.
            Reconciles as: netChange = -produced + received - released + newIntake + increases.
      required:
        - produced
        - received
        - released
        - newIntake
        - netChange
...
    ContractorBalanceMovementDay:
      type: object
      description: One consecutive import pair's per-contractor movement measures.
      properties:
        importId:
          type: integer
          description: ID of the current (newer) import.
        dayKey:
          type: string
          description: YYYY-MM-DD of the current import.
        dayLabel:
          type: string
          description: Human-readable label, e.g. "20 Jul" or "18-20 Jul (2d)".
        prevImportId:
          type: integer
          description: ID of the previous (older) import.
        prevDayKey:
          type: string
          description: YYYY-MM-DD of the previous import.
        isGap:
          type: boolean
          description: True when the two imports are more than one calendar day apart.
        elapsedDays:
          type: integer
          description: Calendar days between the two import dates.
        contractors:
          type: object
          description: Map of contractor name to movement measures. Blank contractor appears as "(Unassigned)".
          additionalProperties:
            $ref: "#/components/schemas/ContractorBalanceMeasures"
      required:
        - importId
        - dayKey
        - dayLabel
        - prevImportId
        - prevDayKey
        - isGap
        - elapsedDays
        - contractors